requirements.txt
azure-identity
azure-functions
azure-keyvault-secrets
pandas
pyodbc
requests
sqlalchemy